# Update VecDeque implementation to use head+len instead of head+tail