@Sp00ph 2022-10-12
Is there any better way to run the tidy tool just for alloc, since I didn't change any other code?