@Sp00ph 2022-10-13
I think I finally got it to work, sorry for all that.