@Sp00ph 2022-10-13
I'm not familiar with the format of these .stderr files, what do I need to do to make it not fail there?