@Sp00ph 2022-10-13
Oh boy, doing that does an LLVM build for me, which my laptop doesn't have the resources for.