@the8472 2022-10-13
set `download-ci-llvm = "if-available"` in config.toml