@Kobzol 2022-10-13
That's fine, there will be many more failures (those are not even all the tests that have to pass :grin: ). Could you share the results of stdlib (or any other) microbenchmarks of the old and the new `VecDeque` on your machine?