@Sp00ph 2022-10-15
Alright, I'd say the code is ready now, just a quick recap of the thread:

The new `VecDeque` implementation passes all its tests, both running normally and under miri. These are stdlib benchmark comparisons of the old and the new implementation on my machine:

```
Old implementation:
test collections::vec_deque::tests::bench_pop_back_100       ... bench:         259 ns/iter (+/- 4)
test collections::vec_deque::tests::bench_pop_front_100      ... bench:         252 ns/iter (+/- 67)
test collections::vec_deque::tests::bench_push_back_100      ... bench:         288 ns/iter (+/- 5)
test collections::vec_deque::tests::bench_push_front_100     ... bench:         291 ns/iter (+/- 388)
test collections::vec_deque::tests::bench_retain_half_10000  ... bench:     574,685 ns/iter (+/- 38,841)
test collections::vec_deque::tests::bench_retain_odd_10000   ... bench:     570,027 ns/iter (+/- 799,114)
test collections::vec_deque::tests::bench_retain_whole_10000 ... bench:     997,678 ns/iter (+/- 508,766)

test vec_deque::bench_extend_bytes                       ... bench:          20 ns/iter (+/- 35)
test vec_deque::bench_extend_chained_bytes               ... bench:       1,619 ns/iter (+/- 45)
test vec_deque::bench_extend_chained_trustedlen          ... bench:       1,676 ns/iter (+/- 58)
test vec_deque::bench_extend_trustedlen                  ... bench:       1,064 ns/iter (+/- 9)
test vec_deque::bench_extend_vec                         ... bench:          99 ns/iter (+/- 155)
test vec_deque::bench_from_array_1000                    ... bench:         489 ns/iter (+/- 246)
test vec_deque::bench_grow_1025                          ... bench:       4,082 ns/iter (+/- 48)
test vec_deque::bench_iter_1000                          ... bench:         638 ns/iter (+/- 1,029)
test vec_deque::bench_mut_iter_1000                      ... bench:         637 ns/iter (+/- 6)
test vec_deque::bench_new                                ... bench:          61 ns/iter (+/- 102)
test vec_deque::bench_try_fold                           ... bench:         375 ns/iter (+/- 651)

custom-bench vec_deque_append 249800 ns/iter

New implementation:
test collections::vec_deque::tests::bench_pop_back_100       ... bench:          89 ns/iter (+/- 108)
test collections::vec_deque::tests::bench_pop_front_100      ... bench:         169 ns/iter (+/- 136)
test collections::vec_deque::tests::bench_push_back_100      ... bench:         207 ns/iter (+/- 3)
test collections::vec_deque::tests::bench_push_front_100     ... bench:         253 ns/iter (+/- 278)
test collections::vec_deque::tests::bench_retain_half_10000  ... bench:     433,255 ns/iter (+/- 4,701)
test collections::vec_deque::tests::bench_retain_odd_10000   ... bench:     436,163 ns/iter (+/- 398,544)
test collections::vec_deque::tests::bench_retain_whole_10000 ... bench:     276,428 ns/iter (+/- 321,348)

test vec_deque::bench_extend_bytes                       ... bench:          17 ns/iter (+/- 23)
test vec_deque::bench_extend_chained_bytes               ... bench:       1,068 ns/iter (+/- 14)
test vec_deque::bench_extend_chained_trustedlen          ... bench:       1,069 ns/iter (+/- 16)
test vec_deque::bench_extend_trustedlen                  ... bench:       1,292 ns/iter (+/- 1,971)
test vec_deque::bench_extend_vec                         ... bench:         128 ns/iter (+/- 169)
test vec_deque::bench_from_array_1000                    ... bench:         539 ns/iter (+/- 846)
test vec_deque::bench_grow_1025                          ... bench:       3,947 ns/iter (+/- 89)
test vec_deque::bench_iter_1000                          ... bench:         426 ns/iter (+/- 626)
test vec_deque::bench_mut_iter_1000                      ... bench:         788 ns/iter (+/- 993)
test vec_deque::bench_new                                ... bench:           2 ns/iter (+/- 0)
test vec_deque::bench_try_fold                           ... bench:         607 ns/iter (+/- 194)

custom-bench vec_deque_append 238050 ns/iter
```

The variance on some of these benchmarks is a bit high on my machine, so I'll also add the results of @Kobzol running the benchmarks on their laptop:
```
master
collections::vec_deque::tests::bench_pop_back_100       ... bench:         187 ns/iter (+/- 5)
test collections::vec_deque::tests::bench_pop_front_100      ... bench:         165 ns/iter (+/- 2)
test collections::vec_deque::tests::bench_push_back_100      ... bench:         165 ns/iter (+/- 6)
test collections::vec_deque::tests::bench_push_front_100     ... bench:         165 ns/iter (+/- 2)
test collections::vec_deque::tests::bench_retain_half_10000  ... bench:     381,092 ns/iter (+/- 5,772)
test collections::vec_deque::tests::bench_retain_odd_10000   ... bench:     383,285 ns/iter (+/- 10,536)
test collections::vec_deque::tests::bench_retain_whole_10000 ... bench:     355,320 ns/iter (+/- 7,691)

test vec_deque::bench_extend_bytes                       ... bench:          10 ns/iter (+/- 0)
test vec_deque::bench_extend_chained_bytes               ... bench:       1,661 ns/iter (+/- 59)
test vec_deque::bench_extend_chained_trustedlen          ... bench:       1,617 ns/iter (+/- 58)
test vec_deque::bench_extend_trustedlen                  ... bench:       1,238 ns/iter (+/- 51)
test vec_deque::bench_extend_vec                         ... bench:          26 ns/iter (+/- 1)
test vec_deque::bench_from_array_1000                    ... bench:         192 ns/iter (+/- 7)
test vec_deque::bench_grow_1025                          ... bench:       1,986 ns/iter (+/- 90)
test vec_deque::bench_iter_1000                          ... bench:         465 ns/iter (+/- 17)
test vec_deque::bench_mut_iter_1000                      ... bench:         467 ns/iter (+/- 10)
test vec_deque::bench_new                                ... bench:          15 ns/iter (+/- 0)
test vec_deque::bench_try_fold                           ... bench:          34 ns/iter (+/- 0)

custom-bench vec_deque_append 311562 ns/iter

This PR
collections::vec_deque::tests::bench_pop_back_100       ... bench:         148 ns/iter (+/- 1)
test collections::vec_deque::tests::bench_pop_front_100      ... bench:         146 ns/iter (+/- 1)
test collections::vec_deque::tests::bench_push_back_100      ... bench:         169 ns/iter (+/- 4)
test collections::vec_deque::tests::bench_push_front_100     ... bench:         194 ns/iter (+/- 4)
test collections::vec_deque::tests::bench_retain_half_10000  ... bench:     255,548 ns/iter (+/- 5,597)
test collections::vec_deque::tests::bench_retain_odd_10000   ... bench:     257,342 ns/iter (+/- 8,091)
test collections::vec_deque::tests::bench_retain_whole_10000 ... bench:     231,784 ns/iter (+/- 6,616)

test vec_deque::bench_extend_bytes                       ... bench:         695 ns/iter (+/- 16) (regression has since been fixed)
test vec_deque::bench_extend_chained_bytes               ... bench:         813 ns/iter (+/- 32)
test vec_deque::bench_extend_chained_trustedlen          ... bench:         735 ns/iter (+/- 30)
test vec_deque::bench_extend_trustedlen                  ... bench:          15 ns/iter (+/- 0)
test vec_deque::bench_extend_vec                         ... bench:          27 ns/iter (+/- 0)
test vec_deque::bench_from_array_1000                    ... bench:         195 ns/iter (+/- 5)
test vec_deque::bench_grow_1025                          ... bench:       2,074 ns/iter (+/- 418)
test vec_deque::bench_iter_1000                          ... bench:         328 ns/iter (+/- 18)
test vec_deque::bench_mut_iter_1000                      ... bench:         327 ns/iter (+/- 5)
test vec_deque::bench_new                                ... bench:           2 ns/iter (+/- 0)
test vec_deque::bench_try_fold                           ... bench:          34 ns/iter (+/- 2)

custom-bench vec_deque_append 144851 ns/iter
```

I've also fuzzed both the old `VecDeque` and this new implementation against [my original `Deque` rewrite](https://github.com/Sp00ph/deque) for a while, and I haven't seen any difference in their behavior, so I'm reasonably confident that they behave the same.

@rustbot ready